+++ server/src/main/kotlin/org/kryptonmc/krypton/packet/out/play/PacketOutAbilities.kt
@@ -9,15 +9,15 @@ class PacketOutAbilities(private val abilities: Abilities) : PlayPacket(0x30) {
     override fun write(buf: ByteBuf) {
         buf.writeByte(abilities.flagsToProtocol())
         buf.writeFloat(abilities.flyingSpeed)
-        buf.writeFloat(abilities.fieldOfViewModifier)
+        buf.writeFloat(abilities.walkSpeed)
     }
 }
 
 private fun Abilities.flagsToProtocol(): Int {
     var flagsByte = 0x0
     if (isInvulnerable) flagsByte += 0x01
-    if (canFly) flagsByte += 0x02
-    if (isFlyingAllowed) flagsByte += 0x04
-    if (isCreativeMode) flagsByte += 0x08
+    if (isFlying) flagsByte += 0x02
+    if (canFly) flagsByte += 0x04
+    if (canInstantlyBuild) flagsByte += 0x08
     return flagsByte
 }