+++ server/src/main/kotlin/org/kryptonmc/krypton/world/data/PlayerDataManager.kt
@@ -2,19 +2,21 @@ package org.kryptonmc.krypton.world.data
 
 import net.kyori.adventure.nbt.*
 import net.kyori.adventure.nbt.BinaryTagIO.Compression.GZIP
+import net.kyori.adventure.text.Component
+import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer
+import org.kryptonmc.krypton.api.entity.Abilities
 import org.kryptonmc.krypton.api.inventory.item.ItemStack
+import org.kryptonmc.krypton.api.inventory.item.ItemType
+import org.kryptonmc.krypton.api.inventory.item.meta.ItemMeta
 import org.kryptonmc.krypton.api.registry.toNamespacedKey
-import org.kryptonmc.krypton.api.space.Vector
 import org.kryptonmc.krypton.api.world.Gamemode
+import org.kryptonmc.krypton.api.world.Location
 import org.kryptonmc.krypton.entity.Attribute
 import org.kryptonmc.krypton.entity.AttributeKey
 import org.kryptonmc.krypton.entity.entities.KryptonPlayer
-import org.kryptonmc.krypton.entity.entities.data.InventoryItem
-import org.kryptonmc.krypton.entity.entities.data.PlayerAbilities
-import org.kryptonmc.krypton.entity.entities.data.PlayerData
 import org.kryptonmc.krypton.entity.memory.EmptyBrain
-import org.kryptonmc.krypton.inventory.KryptonPlayerInventory
 import org.kryptonmc.krypton.serializers.serialize
+import org.kryptonmc.krypton.world.KryptonWorld
 import org.kryptonmc.krypton.world.NBT_DATA_VERSION
 import java.io.File
 import java.io.IOException
@@ -22,27 +24,29 @@ import java.util.*
 
 class PlayerDataManager(private val folder: File) {
 
-    fun load(uuid: UUID): PlayerData? {
-        val playerFile = File(folder, "$uuid.dat")
+    fun loadAndPopulate(world: KryptonWorld, player: KryptonPlayer) {
+        player.world = world
+
+        val playerFile = File(folder, "${player.uuid}.dat")
         if (!playerFile.exists()) {
             playerFile.createNewFile()
-            return null
+            return
         }
 
         val nbt = try {
             BinaryTagIO.unlimitedReader().read(playerFile.toPath(), GZIP)
         } catch (exception: IOException) {
-            return null
+            return
         }
 
-        val attributes = nbt.getList("Attributes").map { attribute ->
+        player.attributes = nbt.getList("Attributes").map { attribute ->
             (attribute as CompoundBinaryTag).let {
                 Attribute(AttributeKey.fromKey(it.getString("Name").toNamespacedKey()), it.getDouble("Base"))
             }
         }.toSet()
 
-        val abilities = nbt.getCompound("abilities").let {
-            PlayerAbilities(
+        player.abilities = nbt.getCompound("abilities").let {
+            Abilities(
                 it.getBoolean("instabuild"),
                 it.getBoolean("mayfly"),
                 it.getBoolean("invulnerable"),
@@ -53,46 +57,22 @@ class PlayerDataManager(private val folder: File) {
             )
         }
 
-        val inventory = nbt.getList("Inventory").map { item ->
+        val inventoryItems = nbt.getList("Inventory").map { item ->
             (item as CompoundBinaryTag).let {
-                InventoryItem(
-                    it.getByte("Slot"),
-                    it.getString("id").toNamespacedKey(),
-                    it.getCompound("tag"),
-                    it.getByte("Count")
-                )
+                ItemStack(ItemType.valueOf(it.getString("id").split(":")[1]), it.getByte("Count").toInt())
             }
         }
+        player.inventory.populate(9, 35, inventoryItems)
 
         val position = nbt.getList("Pos").map { (it as DoubleBinaryTag).value() }
         val rotation = nbt.getList("Rotation").map { (it as FloatBinaryTag).value() }
 
-        return PlayerData(
-            nbt.getShort("SleepTimer"),
-            nbt.getBoolean("SpawnForced"),
-            attributes,
-            nbt.getBoolean("Invulnerable"),
-            nbt.getFloat("AbsorptionAmount"),
-            abilities,
-            nbt.getFloat("FallDistance"),
-            nbt.getShort("Air", 300),
-            Vector(nbt.getInt("SpawnX"), nbt.getInt("SpawnY"), nbt.getInt("SpawnZ")),
-            nbt.getFloat("SpawnAngle"),
-            nbt.getInt("Score"),
-            Vector(position[0], position[1], position[2]),
-            Gamemode.fromId(nbt.getInt("previousPlayerGameType", -1)),
-            Gamemode.fromId(nbt.getInt("playerGameType", 0)) ?: Gamemode.SURVIVAL,
-            nbt.getInt("SelectedItemSlot"),
-            nbt.getShort("HurtTime"),
-            inventory,
-            nbt.getBoolean("FallFlying"),
-            nbt.getString("SpawnDimension").toNamespacedKey(),
-            nbt.getString("Dimension").toNamespacedKey(),
-            nbt.getFloat("Health", 1F),
-            nbt.getBoolean("OnGround"),
-            rotation[1],
-            rotation[0]
-        )
+        player.location = Location(world, position[0], position[1], position[2], rotation[0], rotation[1])
+        player.oldGamemode = Gamemode.fromId(nbt.getInt("previousPlayerGameType", -1))
+        player.gamemode = Gamemode.fromId(nbt.getInt("playerGameType", 0)) ?: Gamemode.SURVIVAL
+        player.inventory.heldSlot = nbt.getInt("SelectedItemSlot")
+        player.dimension = nbt.getString("Dimension").toNamespacedKey()
+        player.isOnGround = nbt.getBoolean("OnGround")
     }
 
     fun save(player: KryptonPlayer) {
@@ -105,6 +85,23 @@ class PlayerDataManager(private val folder: File) {
                 .build()
         }
 
+        val helmet = player.inventory.helmet
+        val chestplate = player.inventory.chestplate
+        val leggings = player.inventory.leggings
+        val boots = player.inventory.boots
+        val offHand = player.inventory.offHand
+
+        val inventory = ((player.inventory.hotbar + player.inventory.main).mapIndexed { index, item ->
+            if (item == null) return@mapIndexed null
+            item.toNBT(index)
+        }.filterNotNull() as MutableList<CompoundBinaryTag>).apply {
+            if (helmet != null) add(helmet.toNBT(103))
+            if (chestplate != null) add(chestplate.toNBT(102))
+            if (leggings != null) add(leggings.toNBT(101))
+            if (boots != null) add(boots.toNBT(100))
+            if (offHand != null) add(offHand.toNBT(-106))
+        }
+
         BinaryTagIO.writer().write(CompoundBinaryTag.builder()
             .put(EmptyBrain.write())
             .putShort("SleepTimer", 0)
@@ -113,12 +110,12 @@ class PlayerDataManager(private val folder: File) {
             .putBoolean("Invulnerable", player.abilities.isInvulnerable)
             .putFloat("AbsorptionAmount", 0F)
             .put("abilities", CompoundBinaryTag.builder()
-                .putFloat("walkSpeed", 0.1F)
-                .putBoolean("instabuild", player.abilities.isCreativeMode)
+                .putFloat("walkSpeed", player.abilities.walkSpeed)
+                .putBoolean("instabuild", player.abilities.canInstantlyBuild)
                 .putBoolean("mayfly", player.abilities.canFly)
                 .putBoolean("invulnerable", player.abilities.isInvulnerable)
-                .putBoolean("mayBuild", player.gamemode != Gamemode.ADVENTURE && player.gamemode != Gamemode.SPECTATOR)
-                .putBoolean("flying", player.isFlying)
+                .putBoolean("mayBuild", player.abilities.canBuild)
+                .putBoolean("flying", player.abilities.isFlying)
                 .putFloat("flySpeed", player.abilities.flyingSpeed)
                 .build())
             .putFloat("FallDistance", 0F)
@@ -134,11 +131,11 @@ class PlayerDataManager(private val folder: File) {
                 .add(DoubleBinaryTag.of(player.location.y))
                 .add(DoubleBinaryTag.of(player.location.z))
                 .build())
-            .putInt("previousPlayerGameType", -1)
+            .putInt("previousPlayerGameType", player.oldGamemode?.ordinal ?: -1)
             .putInt("DataVersion", NBT_DATA_VERSION)
-            .putInt("SelectedItemSlot", 0)
+            .putInt("SelectedItemSlot", player.inventory.heldSlot)
             .putShort("HurtTime", 0)
-            .put("Inventory", ListBinaryTag.of(BinaryTagTypes.COMPOUND, player.inventory.map(InventoryItem::toNBT)))
+            .put("Inventory", ListBinaryTag.from(inventory))
             .putBoolean("FallFlying", false)
             .putInt("playerGameType", player.gamemode.ordinal)
             .putString("SpawnDimension", player.dimension.toString())
@@ -152,3 +149,17 @@ class PlayerDataManager(private val folder: File) {
             .build(), playerFile.toPath(), GZIP)
     }
 }
+
+private fun ItemStack.toNBT(slot: Int) = CompoundBinaryTag.builder()
+    .putByte("Slot", slot.toByte())
+    .putByte("Count", amount.toByte())
+    .apply { if (meta != null) { put("display", meta!!.toNBT()) } }
+    .build()
+
+private fun ItemMeta.toNBT() = CompoundBinaryTag.builder()
+    .apply { if (displayName != null) putString("Name", GsonComponentSerializer.gson().serialize(displayName!!)) }
+    .apply { if (lore.isNotEmpty()) put("Lore", lore.toNBT()) }
+    .build()
+
+private fun List<Component>.toNBT() =
+    ListBinaryTag.of(BinaryTagTypes.STRING, map { StringBinaryTag.of(GsonComponentSerializer.gson().serialize(it)) })