+++ api/src/main/kotlin/org/kryptonmc/krypton/api/inventory/item/ItemStack.kt
@@ -3,7 +3,16 @@ package org.kryptonmc.krypton.api.inventory.item
 import org.kryptonmc.krypton.api.inventory.item.meta.ItemMeta
 
 data class ItemStack(
-    val type: ItemType,
+    val type: Material,
     val amount: Int,
     val meta: ItemMeta? = null
-)
+) {
+
+    companion object {
+
+        /**
+         * An empty item stack.
+         */
+        val EMPTY = ItemStack(Material.AIR, 0)
+    }
+}