+++ server/src/main/kotlin/org/kryptonmc/krypton/entity/Slot.kt
@@ -4,7 +4,7 @@ import net.kyori.adventure.nbt.CompoundBinaryTag
 
 data class Slot(
     val isPresent: Boolean,
-    val itemId: Int = 0,
-    val itemCount: Byte = 0,
+    val id: Int = 0,
+    val count: Byte = 0,
     val nbt: CompoundBinaryTag = CompoundBinaryTag.empty()
 )